adjust patch to auto re-patch unless version already patched

+++ cmd/patch-ytdlp/run.sh
@@ -3,16 +3,21 @@
 
 export BASE="$(dirname "$(readlink -f "$0")")"
 
-YT_DLP_VERSION=${1?:"usage: $0 <version>"}
-
-PATCH_DIR="${BASE}/tmp/${YT_DLP_VERSION}"
+YTDLP_VERSION=${1?:"usage: $0 <version>"}
+PATCH_DIR="${BASE}/tmp/${YTDLP_VERSION}"
+EXPORT_FILE="${BASE}/export-${YTDLP_VERSION}.json"
 
 # if [ -d "$PATCH_DIR" ]; then
 # 	echo "yt-dlp patch already completed for version, not doing anything"
 # 	exit 0
 # fi
 
-echo "patching yt-dlp @ ${YT_DLP_VERSION}"
+if [ -f "$EXPORT_FILE" ]; then
+	echo "yt-dlp export already exists for version at '${EXPORT_FILE}', not doing anything"
+	exit 0
+fi
+
+echo "patching yt-dlp @ ${YTDLP_VERSION}"
 
 if [ -d "$PATCH_DIR" ]; then
 	rm -rf "$PATCH_DIR"
@@ -26,7 +31,7 @@ mkdir -vp "$PATCH_DIR"
 		-c advice.detachedHead=false \
 		clone \
 		--depth 1 \
-		--branch "$YT_DLP_VERSION" \
+		--branch "$YTDLP_VERSION" \
 		https://github.com/yt-dlp/yt-dlp.git "$PATCH_DIR"
 )
 
@@ -39,4 +44,4 @@ if ! grep -q -- "--export-options" "yt_dlp/__main__.py"; then
 	)
 fi
 
-python3 yt_dlp/__main__.py --export-options > "${BASE}/export.json"
+python3 yt_dlp/__main__.py --export-options > "$EXPORT_FILE"
